LCOT correction
Changing some components of the LCOT to be consistent with the Fortran version.

Correcting an error in vehicle lifetimes in China.

Changing when get_lcot is run pre simulation period (was previously running for every region but this is unnecessary given the region loop in the function).

Adding variables to track vehicle purchase price and fuel price component.