Adding taxable fuel categories in FTT-Tr
RTFT (fuel tax) is not technology dependent in the spreadsheets. In the Fortran model, we use an in-code converter to determine what vehicles are subject to the fuel tax. This was only semi-implemented in the python code and it meant that EVs were being subjected to a fuel tax when they shouldn't.

This should fix this for EVs, though I'm not sure what other vehicles should be exempt. Plug-in hybrids, hydrogen, CNG?

Also changed how the fuel cost component is displayed in the front end - now consistent with Fortran version and includes the fuel tax component.