Tidy FTT-Fr
Deleting some old spreadsheets. Tidying the FTT-Fr code. Changing the variable name for vehicle capacity (ZEWG->ZEWK) to be consistent with the other models. ZEWG is now demand by vehicle.

Created a new variable (ZEWI) for new sales - this was previously using ZEWY which, to be consistent with other FTTs, should be new investment. This switch hasn't been done yet and ZEWY is still new sales. ZEWI is not used for now.