Fixing regulations and exog additions
Regulations and exogenous additions were previously in share space. To be consistent with FTT-Tr (and for compatibility with the IEA EV stock data), this is changed to capacity space.

Clarifying ZREG and ZWSA in the spreadsheets.

Also adding new IEA stock data for small and large electric trucks in Belgium (I used this for testing ZWSA and thought it may as well stay in).