Updating FTT_variables.xlsx and removing old masterfiles which cause new ftt version to crash
Old masterfiles are stored in separate folders to prevent error codes until new masterfiles can be created.

A bug still exists if more than one FTT module is run at once, but not if they are run individually. This seems to be something to do with the new masterfile implementation.