Correction to FTT-Tr start of learning
Some regions have end of shares data in 2019 which was causing TEWW to increase between 2020-2022. I don't think this should be happening since the input values for TEWW are for 2022.

This commit makes sure that TEWW only increases after 2022 and learning only takes place after this year too. This is controlled by using TEWW histend.