Correcting FTT-Tr fuel tax implementation
RTFT (fuel tax) was changing the fuel cost component by extreme amounts. This was because of a missing conversion from $/litre (as is in the Masterfile) to $/MJ (needed to calculate fuel tax costs).

Right now, I'm using an approximate conversion of 35MJ/l but this may be inappropriate for long term use given the different energy content of fuels.